feat: add discrete ppo
