+++ src/projects/rl/policy.py
@@ -1,6 +1,5 @@
 import typing
 
-import chex
 import flax.linen as nn
 import jax
 
@@ -9,7 +8,6 @@ import jax
 # this is a sckeleton architecture, we need another implementation for DQN model
 # NOTE: sometimes the outputs are actions, but for DQN, the outputs are Q-values
 # for all actions.
-@chex.dataclass
 class MlpPolicy(nn.Module):
     r"""Multi-layer Perceptron Policy Network.
 
@@ -42,8 +40,6 @@ class MlpPolicy(nn.Module):
 
         Returns:
             For DQN: Raw Q-values for all actions, with shape `(*, out_features)`
-            For PPO: Action probabilities for all actions, with shape 
-                `(*, out_features)`.
         """
         out = inputs.astype(self.dtype)
         for i in range(self.num_layers):
@@ -73,7 +69,6 @@ class MlpPolicy(nn.Module):
         return out
     
 
-@chex.dataclass
 class ActorCriticPolicy(nn.Module):
     r"""Actor-Critic Policy Network.
 
@@ -98,14 +93,13 @@ class ActorCriticPolicy(nn.Module):
     param_dtype: typing.Any = None
 
     @nn.compact
-    def __call__(self, inputs: jax.Array) -> jax.Array:
+    def __call__(self, inputs: jax.Array) -> typing.Tuple[jax.Array, jax.Array]:
         r"""Forward pass the policy network `\pi(a|s;\theta)`
 
         Args:
             inputs (jax.Array): Input state array of shape `(*, D)`
 
         Returns:
-            For DQN: Raw Q-values for all actions, with shape `(*, out_features)`
             For PPO: Action probabilities for all actions, with shape 
                 `(*, out_features)`.
         """
@@ -134,7 +128,7 @@ class ActorCriticPolicy(nn.Module):
                 # apply activation function for hidden layers
                 out = self.activation_fn(out)
 
-        # NOTE: do we need another activation function for the output layer?
+        # Multi-head actor-critic sharing the same backbone network.
         out = self.activation_fn(out)
 
         # Actor Head (logits for action probabilities)
